Move ObjectDP/SliceDP into own fiels and run TPTests at the end of integration tests
